requirements.txt
absl-py==2.1.0
accelerate
transformers
datasets
openai==1.30.1
torch==2.3.0
tiktoken==0.6.0
ruff==0.5.7
scikit-learn==1.5.1
PyYAML==6.0.1
sentence-transformers
